+++ pkg/specgen/specgen_remote.go
@@ -1,0 +1,8 @@
+//go:build remote
+// +build remote
+
+package specgen
+
+// Empty stub we do not use any libimage on the remote client,
+// this drastically decreases binary size for the remote client.
+type cacheLibImage struct{}
